# lib/api-zod/src/generated/types/openaiConversation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LRSTORE LiveChat AI API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: Date;
}